# perf(@angular/cli): enable Node.js compile code cache when available