--- a/packages/angular/cli/bin/bootstrap.js
+++ b/packages/angular/cli/bin/bootstrap.js
@@ -18,4 +18,12 @@
  * range.
  */
 
-import('../lib/init.js');
+// Enable on-disk code caching if available (Node.js 22.8+)
+try {
+  const { enableCompileCache } = require('node:module');
+
+  enableCompileCache?.();
+} catch {}
+
+// Initialize the Angular CLI
+void import('../lib/init.js');